+++ CHANGES.md
@@ -10,6 +10,8 @@
   * Code generation: Create directory path if it does not exist. Thanks to Robert Dockins
     for the patch.
 
+  * Add support for sRealToSInteger. (Essentially the floor function for SReal.)
+
 ### Version 5.12, 2016-06-06
 
   * Fix GHC8.0 compliation issues, and warning clean-up. Thanks to Adam Foltzer for the bulk